requirements.txt
llama-index==0.11.20
llama-index-llms-openai-like==0.2.0
llama-index-embeddings-huggingface==0.3.1
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-vector-stores-postgres==0.2.5
llama-index-vector-stores-chroma==0.2.0
llama-index-retrievers-bm25==0.3.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
chromadb==0.5.20
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
torch==2.5.1
torchvision==0.20.1
transformers==4.46.2
sentence-transformers==3.2.1
FlagEmbedding==1.2.11
easyocr==1.7.1
docling==2.7.2
docling-core==2.5.2
docling-ibm-models==2.0.3
docling-parse==2.0.2
python-multipart==0.0.12
aiofiles==24.1.0
Pillow==10.4.0
openpyxl==3.1.5
python-docx==1.1.2
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.24.0
magic-pdf[full]==0.7.0b1
sqlalchemy==2.0.23
alembic==1.13.1